I fixed the comments from the previous commit but some of them are still not completed. Please review my comments in the previous commit.
